requirements.txt
# Core dependencies for LongDA benchmark
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# LLM agent framework
smolagents>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0

# Document processing and retrieval
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pypdf>=3.0.0

# Optional: Monitoring and tracing
langfuse>=2.0.0
openinference-instrumentation-smolagents>=0.1.0
